% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_default.R
\name{env_default}
\alias{env_default}
\title{Use default value if environment variable is empty}
\usage{
env_default(env, default)
}
\arguments{
\item{env}{Name of the environment variable as string}

\item{default}{Default value to use when environment variable is empty}
}
\value{
Default value or content of environment variable
}
\description{
Use default value if environment variable is empty
}
\examples{
locale <- env_default(env = "locale", default = "e_DE.utf8")
}
